backend/requirements.txt
pip~=24.1.1
Jinja2~=3.1.4
Flask~=3.0.3
python-dotenv~=1.0.1
pymongo~=4.8.0
attrs~=23.2.0
aiohttp~=3.9.5
Flask-Cors~=4.0.1
Flask-Limiter~=3.7.0
asgiref~=3.8.1
gunicorn~=22.0.0